web/requirements.txt

# Core Django
django==5.1.4
channels>=4.0
daphne>=4.0
channels-redis>=4.0
django-webpush
whitenoise>=6.6.0

# AI
google-generativeai

# Existing
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
dnspython==2.8.0
docker==7.1.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
fonttools==4.57.0
future==1.0.0
h11==0.16.0
idna==3.11
iso8601==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.3
matplotlib==3.10.1
numpy==2.2.5
packaging==25.0
passlib==1.7.4
pillow==11.2.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
scikit-fuzzy==0.5.0
scipy==1.15.2
simpy==4.1.1
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
gunicorn
